+ use long instead of int + add dome messages to Controller + fix some issues

+++ Zynarator/Controller/BaseController.cs
@@ -4,6 +4,7 @@ using DotnetGenerator.Zynarator.Dto;
 using DotnetGenerator.Zynarator.Service;
 using Lamar;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.Tokens;
 
 namespace DotnetGenerator.Zynarator.Controller;
 
@@ -17,40 +18,48 @@ public abstract class BaseController<TEntity, TDto, TService> : ControllerConver
     protected BaseController(IServiceContext container, IMapper mapper) : base(mapper) =>
         Service = container.GetInstance<TService>();
     
-    public virtual async Task<ActionResult<List<TDto>>> FindAll() =>
-        Ok(ToDto(await Service.FindAll()));
-    
+    public virtual async Task<ActionResult<List<TDto>>> FindAll()
+    {
+        var found = await Service.FindAll();
+        if (found.IsNullOrEmpty()) NotFound("No Data Found!");
+        return Ok(ToDto(found));
+    }
+
     public virtual async Task<ActionResult<TDto>> Create(TDto dto)
     {
         var item = ToItem(dto);
-        return Ok(await Service.Create(item));
+        if (item == null) return BadRequest("No Entity Was Provided!");
+        return Created("New Item Was Created Successfully!", await Service.Create(item));
     }
 
     public virtual async Task<ActionResult<List<TDto>>> Create(List<TDto> dtos)
     {
+        if (dtos.IsNullOrEmpty()) return BadRequest("no data was provide to be created!");
         var items = ToItem(dtos);
-        return Ok(await Service.Create(items));
+        return Created("The List Was Created Successfully!", await Service.Create(items));
     }
 
     public virtual async Task<ActionResult<TDto>> Update(TDto dto)
     {
         var item = ToItem(dto);
+        if (item == null) return BadRequest("No Entity Was Provided!");
         return Ok(await Service.Update(item));
     }
 
     public virtual async Task<ActionResult<List<TDto>>> Update(List<TDto> dtos)
     {
+        if (dtos.IsNullOrEmpty()) return BadRequest("no data was provide to be updated!");
         var items = ToItem(dtos);
-        return items.Count == 0 ? Conflict(dtos) : Ok(await Service.Update(items));
+        return Ok(await Service.Update(items));
     }
     
-    public virtual async Task<ActionResult<TDto>> FindById(int id)
+    public virtual async Task<ActionResult<TDto>> FindById(long id)
     {
-        var result = await Service.FindById(id);
-        return Ok(ToDto(result));
+        var found = await Service.FindById(id);
+        return Ok(ToDto(found));
     }
 
-    public virtual async Task<ActionResult<int>> DeleteById(int id)
+    public virtual async Task<ActionResult<int>> DeleteById(long id)
     {
         var result = await Service.DeleteById(id);
         return Ok(result);
@@ -59,12 +68,14 @@ public abstract class BaseController<TEntity, TDto, TService> : ControllerConver
     public virtual async Task<ActionResult<int>> Delete(TDto dto)
     {
         var item = ToItem(dto);
+        if (item is null) return BadRequest("no data was provide to be deleted!");
         var result = await Service.Delete(item);
         return Ok(result);
     }
 
     public virtual async Task<ActionResult<int>> Delete(List<TDto> dtos)
     {
+        if (dtos.IsNullOrEmpty()) return BadRequest("no data was provide to be deleted!");
         var items = ToItem(dtos);
         var result = await Service.Delete(items);
         return Ok(result);
